Fix small bug with application configuration

+++ src/main/java/to/us/resume_builder/business/ApplicationConfiguration.java
@@ -100,7 +100,7 @@ public class ApplicationConfiguration {
      */
     public Long getLong(String key) {
         try {
-            return Long.parseLong(configuration.get(key).toString());
+            return (long) Double.parseDouble(configuration.get(key).toString());
         } catch (Exception e) {
             return null;
         }
